[🛠️ REFACTOR] 게임 방 구독 관련 함수 모듈화
